Add some shapes and shadows

+++ src/geometry/AGeomerty.cpp
@@ -1,5 +1,5 @@
 #include "geometry/AGeomerty.hpp"
 
-AGeomerty::AGeomerty() : color({static_cast<uint8_t>(0), static_cast<uint8_t>(0), static_cast<uint8_t>(0)}) { }
+AGeomerty::AGeomerty() : color({0.0, 0.0, 0.0}) { }
 
 AGeomerty::AGeomerty(Color &color_) : color(color_) { }
